v1_vision/requirements.txt
# V1 Vision Pipeline Requirements

# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.0

# NEST Simulator (spiking neural network simulator)
# Note: NEST may require system-level dependencies
# See: https://nest-simulator.readthedocs.io/en/stable/installation/index.html
nest-simulator>=3.0

# Optional: For enhanced visualization
matplotlib>=3.3.0

# Note: The LIFL_IE NEST module must be compiled separately
# See: ../MDPI2021/LIFL_IE/README for installation instructions

